feat: featch backend api (send request)

+++ customer_app/lib/data/models/requests/create_triprequest_request.dart
@@ -18,7 +18,7 @@ class CreateTripRequestBody {
   String PassengerNote;
   double Distance;
   String PassengerPhone;
-  double Price;
+  int Price;
   String VehicleType;
 
   CreateTripRequestBody(
